manager/requirements.txt
docker-py
flask
gunicorn
redis
rq
